word game upload
This	program	involves one turn in a simple word game based on one from the New York Times In the game. A player tries to make as many valid words as possible from a list of letters to get the most points possible. This program also uses Input/Output imports and Java Utility imports such as hash sets, ArrayLists, FileReader, BufferedReader, and scanner for user input.